### 2024-10-10 #7342 @editorialbot
[PRE REVIEW]: popclass: a python package for classifying microlensing events
**Submitting author:** @astrophpeter (<a href="http://orcid.org/0000-0002-1052-6749">Peter McGill</a>)
**Repository:** https://github.com/LLNL/popclass
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
