### 2024-10-10 #7342 @warrickball
In the meantime, if you could try to fix the errors in the paper. If Caitlin Begbie doesn't have an ORCID, you can just remove that line from their metadata.

As indicated, the DOI error looks like it's because the full URL (including `https://doi.org/`) has been given for Greenfield et al. (2015). You should replace this with just the DOI.

Finally, note that your paper is a bit long for JOSS. As mentioned in [the docs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain),

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

If you could squeeze it back under 1000 words, that'd be great.
