+++ JOSS/paper.md
@@ -28,7 +28,7 @@ authors:
     affiliation: 1
     equal-contrib: false
   - name: Caitlin G. Begbie
-    orcid: 0000-0000-0000-0000
+    orcid: 0009-0006-8866-4224
     affiliation: 3
 
 affiliations:
