### 2024-10-10 e376d8e
trimmed paper
