+++ JOSS/paper.md
@@ -42,8 +42,6 @@ affiliations:
 date: 13 August 2017
 bibliography: paper.bib
 
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
 aas-doi:
 aas-journal:
 ---
@@ -52,31 +50,30 @@ aas-journal:
 
 `popclass` is a python package that provides a flexible, probabilistic framework for classifying
 the lens of a gravitational microlensing event. Gravitational microlensing occurs when a massive
-foreground object - the lens (e.g., a star, white dwarf or a black hole) passes in front of and
-deflects the light from a distant background star. This causes an apparent brightening, and shift
+foreground object (e.g., a star, white dwarf or a black hole) passes in front of and
+lenses the light from a distant background star. This causes an apparent brightening, and shift
 in position, of the background source. In most cases, characteristics of the microlensing signal
 do not contain enough information to definitively identity the lens type. However, different lens
-types (e.g., stars vs black holes) can lie in different but overlapping regions of the characteristics
-of the microlensing signal. For examples, black holes tend to be more massive than stars and
-therefore cause microlensing signals that are longer. Current Galactic model simulations allow
+types can lie in different but overlapping regions of the characteristics
+of the microlensing signal. For example, black holes tend to be more massive than stars and
+therefore cause microlensing signals that are longer. Current Galactic simulations allow
 the prediction of where different lens types lie in the observational space and can therefore be
- leveraged to classify an event [e.g., @Lam2020].
+leveraged to classify events [e.g., @Lam2020].
 
 `popclass` allows a user to match characteristics of a microlensing signal to a simulation of the
-Galaxy to calculate lens type probabilities for an event. The user can flexibly use constraints on
-any microlensing signal characteristics and specify their own Galactic model. `popclass` also
+Galaxy to calculate lens type probabilities for an event (see Figure 1). Constraints on
+any microlensing signal characteristics and any Galactic model can be used. `popclass`
 comes with an interface to `ArviZ` [@arviz_2019] and `pymultinest` [@Buchner2016] for microlensing
-signal constraints, pre-loaded Galactic models, plotting functionality, and uncertainty
-quantification methods that can be included in the classification calculation. The probabilistic
-framework for popclass was developed in @Perkins2024, used in @Fardeen2024
-and has been applied to classify microlensing events in @Kaczmarek2024.
+signal constraints, pre-loaded Galactic models, plotting functionality, and classification uncertainty
+quantification methods. The probabilistic framework for popclass was developed in @Perkins2024, 
+used in @Fardeen2024 and has been applied to classifying events in @Kaczmarek2024.
 
 # Statement of need
 
 The advent of the _Vera C. Rubin Observatory_ [@Ivezic2019] and the _Nancy Grace Roman Space Telescope_ [@Spergel2015]
-will trigger a deluge of tens-of-thousands of microlensing events per year [e.g., @Abrams2023;@Penny2019]. To maximize the science output of this event stream it is critical to identify events that have a high probability of being caused
-by interesting lens types such as an isolated black hole [@Sahu2022; @Lam2022], and then to allocate expensive
-follow-up observations such as space-based astrometry [e.g., @Sahu2022] or ground-based adaptive optics imaging [e.g., @Terry2022]  to confirm their nature.
[…]
