### 2024-10-10 4b9987f
trimmed paper more
