+++ .pre-commit-config.yaml
@@ -1,17 +1,17 @@
 exclude: 'popclass/data/.*|examples/.*'
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v4.6.0
+  rev: v5.0.0
   hooks:
   - id: check-yaml
   - id: end-of-file-fixer
   - id: trailing-whitespace
   - id: check-byte-order-marker
 - repo: https://github.com/psf/black
-  rev: 23.12.1
+  rev: 24.10.0
   hooks:
   - id: black
 - repo: https://github.com/asottile/reorder-python-imports
-  rev: v3.12.0
+  rev: v3.14.0
   hooks:
   -   id: reorder-python-imports
