### 2024-10-17 d71735e
minor edits in paper.md
fixed several typos, edited one sentence that had a changed subject
