### 2025-02-10 #7769 @editorialbot
[REVIEW]: popclass: A Python Package for Classifying Microlensing Events
**Submitting author:** @astrophpeter (<a href="http://orcid.org/0000-0002-1052-6749">Peter McGill</a>)
**Repository:** https://github.com/LLNL/popclass
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.1
