### 2025-02-10 #7769 @ivastar
@skterry, @ebachelet welcome to the review issue for `popclass`!

The instructions for generating the reviewer checklist are above. For more details on each of the items we ask you to address in the review, please refer to the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The timeline for the initial review is ~3 weeks, let's aim for March 3rd, but let me know if you need more time.

Don't hesitate to reach out here or via email if you have any questions or concerns.
