### 2025-02-24 #7769 @skterry
Additional details of my review:

The code is quite clear and the paper is well written, with a new classifier for long duration microlensing events that is certainly worth publishing. I just have a few additional comments and/or suggestions regarding the code.

(1) There is some ambiguity in readthedocs documentation regarding the ‘/data’ directory. There are two ‘/data’ directories:
popclass/popclass/data/
	and
popclass/docs/source/examples/data/

For example, at this page: https://popclass.readthedocs.io/en/latest/examples/popclass_example_workflow.html

Says the data is located in /data/parallax_trace.csv ‘within this parent directory’. Some users may only follow readthedocs and not use the Jupyter Notebook. I simply suggest using the full path for the example data (i.e. popclass/docs/source/examples/data/parallax_trace.csv).

(2) Inside test/test_classify.py,
The function ‘test_BH_example()’ includes a line to ‘print(classification)’, but the other similar test functions (e.g. ‘test_star_example()’ and ‘test_full_example()’) do not include this print statement. I suggest either adding a print statement to each of these test functions, or removing the print statement from the test_BH_example() function.

(3) The last few cells of the ‘Working with posteriors and popclass’ tutorial Notebook are producing a negative posterior probability for parameter ‘A’. Please check that this is the intended functionality and result that is expected. If not, please correct.

-------------
Lastly, it is beyond the scope of this review of the code and its functionality, however I tested the popclass classifier on the first confirmed isolated black hole from gravitational microlensing, OGLE-2011-BLG-0462 (Sahu+2022, Lam & Lu 2023).

Using Table 5 of Sahu+2022 parameters for log_tE and log_piE, popclass gives the following probabilities:
{'black_hole': 0.2211442623279775, 'neutron_star': 3.2993436756897125e-07, 'star': 0.3416396702174749, 'white_dwarf': 0.43721573752018006}

And using Table 4 of Lam & Lu (2023) parameters for log_tE and log_piE, popclass gives the following probabilities:
{'black_hole': 0.13720552085293217, 'neutron_star': 3.3498504658534105e-08, 'star': 0.4207216505040563, 'white_dwarf': 0.44207279514450676}
