+++ tests/test_classify.py
@@ -72,8 +72,6 @@ def test_BH_example():
         population_model=popsycle, inference_data=inference_data, parameters=parameters
     )
 
-    print(classification)
-
     assert abs(1.0 - classification["black_hole"]) < 0.0001
 
 
