+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-sphinx
+sphinx==7.2.3
 sphinx_rtd_theme
 sphinxcontrib-bibtex
 nbsphinx
