### 2025-03-11 #7769 @ivastar
> Sorry Ivelina, I totally forgot about additional comments. I thought I did
> everything. It is good to go for me.
> I will keep this in mind for the next one as I am busy right now.
>
> Best
> […](#)

Just one item is left unchecked and I thought you maybe had a comment about it:
"Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support"

If you don't have any comments about this, I would appreciate you checking off the item in the checklist so it doesn't look like there is anything that needs to be addressed. Thanks!
