### 2025-03-11 #7769 @ebachelet
I may have missed it but I do not see on the repository clear guidelines about:
"Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support"

So I left this uncommented. I would suggest the author to add such sentences, this is trivial.
