+++ JOSS/paper.bib
@@ -226,13 +226,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@unpublished{Kaczmarek2024,
-       author = {{Kaczmarek}, Z. and {McGill}, P. and {Perkins}, S. and {Dawson}, W. and {Abrams}, N. and {Ho}, M. and {Huston}, M. and {Lu}, J.},
+@ARTICLE{Kaczmarek2024,
+       author = {{Kaczmarek}, Zofia and {McGill}, Peter and {Perkins}, Scott E. and {Dawson}, William A. and {Huston}, Macy and {Ho}, Ming-Feng and {Abrams}, Natasha S. and {Lu}, Jessica R.},
         title = "{On Finding Black Holes in Photometric Microlensing Surveys}",
-         year = 2024,
-        month = sep,
-	note="in preparation"
+      journal = {\apj},
+     keywords = {Gravitational microlensing, Black holes, Bayesian statistics, Classification, Sky surveys, 672, 162, 1900, 1907, 1464, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2025,
+        month = mar,
+       volume = {981},
+       number = {2},
+          eid = {183},
+        pages = {183},
+          doi = {10.3847/1538-4357/adb1d7},
+archivePrefix = {arXiv},
+       eprint = {2410.14098},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025ApJ...981..183K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @ARTICLE{Ivezic2019,
[…]
