+++ JOSS/paper.bib
@@ -328,7 +328,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Howil2025,
+@ARTICLE{Howil2024,
        author = {{Howil}, K. and {Wyrzykowski}, {\L}. and {Kruszy{\'n}ska}, K. and {Zieli{\'n}ski}, P. and {Bachelet}, E. and {Gromadzki}, M. and {Miko{\l}ajczyk}, P.~J. and {Kotysz}, K. and {Jab{\l}o{\'n}ska}, M. and {Kaczmarek}, Z. and {Mr{\'o}z}, P. and {Ihanec}, N. and {Ratajczak}, M. and {Pylypenko}, U. and {Rybicki}, K. and {Sweeney}, D. and {Hodgkin}, S.~T. and {Larma}, M. and {Carrasco}, J.~M. and {Burgaz}, U. and {Godunova}, V. and {Simon}, A. and {Cusano}, F. and {Jelinek}, M. and {{\v{S}}trobl}, J. and {Hudec}, R. and {Merc}, J. and {Ku{\v{c}}{\'a}kov{\'a}}, H. and {Erece}, O. and {Kilic}, Y. and {Olivares}, F. and {Morrell}, M. and {Wicker}, M.},
         title = "{Uncovering the invisible: A study of Gaia18ajz, a candidate black hole revealed by microlensing}",
       journal = {\aap},
