### 2025-03-20 #7769 @astrophpeter
Hi Everyone,

Thank you for your reviews.

I think I have address all issues in PR https://github.com/LLNL/popclass/pull/74

@skterry (1) Agreed, I made the changes to the file paths in the examples (2) Agreed, the print statement has been removed. (3) I can confirm that this is the intended behavior - parameter A has roughly half of it values being < 0. Other comment, yes that is correct.

@ebachelet I've added a sentence to the document home page pointing users to Github issues for help and support. There is also a lot of details in the contributing guide for developers https://popclass.readthedocs.io/en/latest/contributing.html

I also updated some references in paper.bib.

@ivastar I think everything is completed on my end.
