### 2025-05-05 #7769 @ivastar
Some minor editorial comments:

Line 1: please capitalize title appropriately
Line 10: Python should be capitalized
Line 13: are the sources always stars? would it be more appropriate to just use "source"?
Line 14: In most cases characteristics --> In most cases, the characteristics
Line 15: However, different lens types --> Different
Line 16: perhaps remove "can", they do lie in different regions, yes?
Line 16: characteristics --> parameter space?
Line 18: allow the prediction of --> enable us to predict (allow repeats in the next sentence)
Line 19: leveraged --> used
Line 21: allows a user to match characteristics --> allows the user to match the characteristics
Line 21: signal to a --> signal with a
Line 22: an event --> the event
Line 23: characteristics -->  properties
Line 25: constraints, pre-loaded --> constraints, as well as pre-loaded
Line 26: classification uncertaintly quantification methods --> methods to quantify the classification uncertainty
Line 31: tens of thousands doesn't nee hypthens
Line 32: science output of  --> scientific output from
Line 33: a comma after "stream"
Line 34: by a lens type --> by a certain lens type?
Line 34: "and then to allocate" --> maybe "before we allocate"
Line 37: mass-distance --> mass and distance?
Line 37/38: using microlensing even light curve --> using the light curves of microlensing events?
Line 38: remove "additional" (redundant with "auxiliary"
Line 41: limits them to only be effective for --> limits their application to
Line 42: predict lens type --> predict the lens type
Line 48: relies on --> is based on
Line 52: the classes should not be capitalized: star, neutron star, etc.
Line 63: is just set by relative --> is simply set by the relative

Some notes on the references:
- Please update the Abrams citation to the ApJS one: https://ui.adsabs.harvard.edu/abs/2025ApJS..276...10A/abstract

- The Buchner citation for pymultinest should be to this paper: https://johannesbuchner.github.io/PyMultiNest/index.html#citing-pymultinest

- The main scikit-learn citation is Pedregosa, I suggest adding this to the Buitinick: https://scikit-learn.org/0.21/about.html#citing-scikit-learn

- Please replace all shortcuts for journal names (e.g., \apj) with the name journal because out system does not recognize these

Following these changes, please proceed to the items below.
