### 2025-05-14 4878eeb
updated paper and bib for joss review
