+++ JOSS/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'popclass: a python package for classifying microlensing events'
+title: 'popclass: A Python Package for Classifying Microlensing Events'
 tags:
   - Python
   - astronomy
@@ -48,49 +48,49 @@ aas-journal:
 
 # Summary
 
-`popclass` is a python package that provides a flexible, probabilistic framework for classifying
+`popclass` is a Python package that provides a flexible, probabilistic framework for classifying
 the lens of a gravitational microlensing event. Gravitational microlensing occurs when a massive
 foreground object (e.g., a star, white dwarf or a black hole) passes in front of and
-lenses the light from a distant background star. This causes an apparent brightening, and shift
+lenses the light from a distant background source. This causes an apparent brightening, and shift
 in position, of the background source. In most cases, characteristics of the microlensing signal
-do not contain enough information to definitively identify the lens type. However, different lens
-types can lie in different but overlapping regions of the characteristics
+do not contain enough information to definitively identify the lens type. Different lens
+types lie in different but overlapping regions of the characteristics
 of the microlensing signal. For example, black holes tend to be more massive than stars and
-therefore cause microlensing signals that are longer. Current Galactic simulations allow
-the prediction of where different lens types lie in the observational space and can therefore be
-leveraged to classify events [e.g., @Lam2020].
+therefore cause microlensing signals that are longer. Current Galactic simulations enable us to predict 
+where different lens types lie in the observational space and can therefore be
+used to classify events [e.g., @Lam2020].
 
-`popclass` allows a user to match characteristics of a microlensing signal to a simulation of the
-Galaxy to calculate lens type probabilities for an event (see Figure 1). Constraints on
-any microlensing signal characteristics and any Galactic model can be used. `popclass`
-comes with an interface to `ArviZ` [@arviz_2019] and `pymultinest` [@Buchner2016] for microlensing
-signal constraints, pre-loaded Galactic models, plotting functionality, and classification uncertainty
-quantification methods. The probabilistic framework for popclass was developed in @Perkins2024,
+`popclass` allows the user to match the characteristics of a microlensing signal with a simulation of the
+Galaxy to calculate lens type probabilities for the event (see Figure 1). Constraints on
+any microlensing signal properties and any Galactic model can be used. `popclass`
+comes with an interface to `ArviZ` [@arviz_2019] and `pymultinest` [@Buchner2014] for microlensing
+signal constraints, as well as pre-loaded Galactic models, plotting functionality, and methods to quantify the classification uncertainty.
+The probabilistic framework for popclass was developed in @Perkins2024,
 used in @Fardeen2024 and has been applied to classifying events in @Kaczmarek2024.
 
 # Statement of need
 
 The advent of the Vera C. Rubin Observatory [@Ivezic2019] and the _Nancy Grace Roman Space Telescope_ [@Spergel2015]
-will provide tens-of-thousands of microlensing events per year [e.g., @Abrams2023;@Penny2019]. To maximize the science output of this event stream it is critical to identify events that have a high probability of being caused
-by a lens type such as a black hole [@Sahu2022; @Lam2022], and then to allocate expensive
+will provide tens of thousands of microlensing events per year [e.g., @Abrams2023;@Penny2019]. To maximize the scientific output from this event stream, it is critical to identify events that have a high probability of being caused
+by a certain lens type such as a black hole [@Sahu2022; @Lam2022], before we allocate expensive
 follow-up observations such as space-based astrometry [e.g., @Sahu2022] or ground-based adaptive optics imaging [e.g., @Terry2022] to confirm their nature.
 
[…]
