+++ JOSS/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Fardeen2024,
        author = {{Fardeen}, James and {McGill}, Peter and {Perkins}, Scott E. and {Dawson}, William A. and {Abrams}, Natasha S. and {Lu}, Jessica R. and {Ho}, Ming-Feng and {Bird}, Simeon},
         title = "{Astrometric Microlensing by Primordial Black Holes with the Roman Space Telescope}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Gravitational microlensing, Astrometry, Primordial black holes, 672, 80, 1292, Astrophysics - Astrophysics of Galaxies},
          year = 2024,
         month = apr,
@@ -17,17 +17,16 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@inproceedings{sklearn_api,
-  author    = {Lars Buitinck and Gilles Louppe and Mathieu Blondel and
-                Fabian Pedregosa and Andreas Mueller and Olivier Grisel and
-                Vlad Niculae and Peter Prettenhofer and Alexandre Gramfort
-                and Jaques Grobler and Robert Layton and Jake VanderPlas and
-                Arnaud Joly and Brian Holt and Ga{\"{e}}l Varoquaux},
-  title     = {{API} design for machine learning software: experiences from the scikit-learn
-                project},
-  booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
-  year      = {2013},
-  pages = {108--122},
+@article{sklearn_api,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
 }
 
 @Article{Hunter2007,
@@ -85,7 +84,7 @@ abstract = {We present the case for developing a successor format for the immens
 @ARTICLE{Terry2022,
        author = {{Terry}, Sean K. and {Bennett}, David P. and {Bhattacharya}, Aparna and {Koshimoto}, Naoki and {Beaulieu}, Jean-Philippe and {Blackman}, Joshua W. and {Bond}, Ian A. and {Cole}, Andrew A. and {Lu}, Jessica R. and {Marquette}, Jean Baptiste and {Ranc}, Cl{\'e}ment and {Rektsini}, Natalia and {Vandorou}, Aikaterini},
         title = "{Adaptive Optics Imaging Can Break the Central Caustic Cusp Approach Degeneracy in High-magnification Microlensing Events}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
[…]
