### 2025-05-15 #7769 @astrophpeter
Thanks @ivastar

I have made all of your suggested changes to the paper.
I have doubled checked all the authors and affiliations.
The new release of the software is version: v0.2.1
The Zenodo DOI of this release is: 10.5281/zenodo.15420667 https://zenodo.org/records/15420667
Authors info and title matches in the Zenodo.
The archive license is the same as the software license.

Please let me know if I missed anything.
