### 2025-05-15 #7769 @ivastar
@astrophpeter seems like the compilation stumbled because there are two definitions of the Greenfield2015 ASDF in the bib file, on line 47 and on line 216. They seem identical. Can you please remove one of them? Also, if there are any other duplicated or unused citation, remove them as well.
