### 2025-05-15 #7769 @astrophpeter
Thanks! The duplicated bib item has now been removed.
