### 2025-05-19 #7769 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
