### 2025-05-23 4cfbdce
Initial editorial tweaks for JOSS paper
