+++ JOSS/paper.md
@@ -3,7 +3,7 @@ title: 'popclass: A Python Package for Classifying Microlensing Events'
 tags:
   - Python
   - astronomy
-  - milky way
+  - Milky Way
   - black hole
 authors:
   - name: Greg Sallaberry
@@ -34,23 +34,21 @@ authors:
 affiliations:
  - name: Space Science Institute, Lawrence Livermore National Laboratory, 7000 East Ave., Livermore, CA 94550, USA
    index: 1
- - name: Zentrum für Astronomie der Universität Heidelberg, Astronomisches Rechen-Institut, Mönchhofstr. 12-14, 69120 Heidelberg, Germany
+ - name: Astronomisches Rechen-Institut, Zentrum für Astronomie der Universität Heidelberg, Mönchhofstr. 12-14, 69120 Heidelberg, Germany
    index: 2
- - name: University of California, Berkeley, Astronomy Department, Berkeley, CA 94720, USA
+ - name: Department of Astronomy, University of California, Berkeley, 501 Campbell Hall, Berkeley, CA 94720, USA
    index: 3
 
 date: 17 October 2024
 bibliography: paper.bib
 
-aas-doi:
-aas-journal:
 ---
 
 # Summary
 
 `popclass` is a Python package that provides a flexible, probabilistic framework for classifying
 the lens of a gravitational microlensing event. Gravitational microlensing occurs when a massive
-foreground object (e.g., a star, white dwarf or a black hole) passes in front of and
+foreground object (e.g., a star, white dwarf or black hole) passes in front of and
 lenses the light from a distant background source. This causes an apparent brightening, and shift
 in position, of the background source. In most cases, characteristics of the microlensing signal
 do not contain enough information to definitively identify the lens type. Different lens
@@ -61,9 +59,9 @@ where different lens types lie in the observational space and can therefore be
 used to classify events [e.g., @Lam2020].
 
 `popclass` allows the user to match the characteristics of a microlensing signal with a simulation of the
-Galaxy to calculate lens type probabilities for the event (see Figure 1). Constraints on
+Galaxy to calculate lens type probabilities for the event (see \autoref{lensclass}). Constraints on
 any microlensing signal properties and any Galactic model can be used. `popclass`
-comes with an interface to `ArviZ` [@arviz_2019] and `pymultinest` [@Buchner2014] for microlensing
+comes with an interface to ArviZ [@arviz_2019] and PyMultiNest [@Buchner2014] for microlensing
 signal constraints, as well as pre-loaded Galactic models, plotting functionality, and methods to quantify the classification uncertainty.
 The probabilistic framework for popclass was developed in @Perkins2024,
 used in @Fardeen2024 and has been applied to classifying events in @Kaczmarek2024.
@@ -75,7 +73,9 @@ will provide tens of thousands of microlensing events per year [e.g., @Abrams202
 by a certain lens type such as a black hole [@Sahu2022; @Lam2022], before we allocate expensive
 follow-up observations such as space-based astrometry [e.g., @Sahu2022] or ground-based adaptive optics imaging [e.g., @Terry2022] to confirm their nature.
 
-Current microlensing software packages such as `DarkLensCode` [@Howil2024] or `PyLiMASS` [@Bachelet2024] estimate
+![Left: posterior distribution of an event in log10(timescale)-log10(parallax) space, overlaid on 'star', 'white dwarf', 'neutron star' and 'black hole' contours. Right: bars showing probabilities of that event belonging to each of the lens populations.\label{lensclass}](lens_class.png)
+
[…]
