### 2025-05-24 #7769 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
